+++ docs/en/changes/changes.md
@@ -2,14 +2,12 @@
 
 #### Project
 
-
 #### OAP Server
 
+* Add `ServerStatusService` in the core module to provide a new way to expose booting status to other modules.
 
 #### UI
 
-
 #### Documentation
-
 
 All issues and pull requests are [here](https://github.com/apache/skywalking/milestone/160?closed=1)